requirements.txt
# Requirements.txt

webbrowser==0.2.1
python-dotenv==1.0.1
openai==1.12.0
termcolor==2.4.0
requests==2.31.0

